pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "fuzzywuzzy>=0.18.0",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-levenshtein>=0.27.3",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
]
